Upates to signup list and the daily

+++ attendees-delete.php
@@ -1,7 +1,10 @@
 <?php
 
-	/* Template Name: DELETE Attendee */
+	// Template Name: DELETE Attendee
 	// Create page with this file as the template
+
+	$path = $_SERVER['DOCUMENT_ROOT']; 
+	include_once $path . '/wp-load.php';
 
 	$form_data = json_decode(file_get_contents("php://input"));
 
@@ -9,38 +12,43 @@
 
 	$database = 'wp_service_attendees';
 	$currentTime = current_time('mysql');
+	$data = array('expired' => $currentTime);
 
-	if($form_data === null){
+	if($_SERVER['REQUEST_METHOD'] === 'POST') {
+		if($form_data === null){
 
-		$get_ids = $wpdb->get_col("
-			SELECT id
-			FROM $database
-			WHERE (`id` > 1 AND `expired` IS NULL)
-		");
-	
-		foreach($get_ids as $key => $value){
-			$data = array('expired' => $currentTime);
-			$where = array('id' => $value);
+			$get_ids = $wpdb->get_col("
+				SELECT id
+				FROM $database
+				WHERE (`id` > 1 AND `expired` IS NULL)
+			");
+		
+			foreach($get_ids as $value){
+				$where = array('id' => $value);
 
-			$wpdb->update(
+				$wpdb->update(
+					$database,
+					$data,
+					$where,
+				);
+			};
+
+		 	echo "Expired";	
+		}
+
+		else {
+			$wpdb->delete(
 				$database,
-				$data,
-				$where,
+				array(
+					'id' => $form_data
+					)
 			);
 
-			print_r( $where );
-		}		
+			echo "Deleted";
+		}
 	}
 
 	else {
-  		$wpdb->delete(
-			$database,
-			array(
-				'id' => $form_data
-				)
-		);
-	}
-
-	echo $ids;
-
+		echo "Don't access this file directly";
+	};
 ?>
